adjust strcuture
